% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchdata-data.R
\docType{data}
\name{catchdata}
\alias{catchdata}
\title{Landings data example}
\format{
An object of class \code{"data.frame"}
}
\source{
Private data
}
\usage{
data(catchdata)
}
\description{
All datasets that contain landing information should follow this format. Be
careful with column names, they should be exactly as shown in this example.
}
\examples{
data(catchdata)
}
\references{
None
}
\keyword{datasets}
