% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distCoast.R
\name{distCoast}
\alias{distCoast}
\title{Distance to the shoreline}
\usage{
distCoast(lon, lat, mainLand = TRUE)
}
\arguments{
\item{lon}{Longitude vector in decimal units (e.g. -76.5242).}

\item{lat}{Latitude vector in decimal units (e.g. -12.1352).}

\item{mainLand}{Logical value. Indicates if distance is calculated
to the shoreline (TRUE) or to shoreline and islands (FALSE).}
}
\value{
A vector with distances (in nautical miles).
}
\description{
Calculates minimum distance to the Peruvian shoreline.
}
