% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icollocation.R
\name{get_icollocation}
\alias{get_icollocation}
\title{Global index of collocation}
\usage{
get_icollocation(data, vessel_name)
}
\arguments{
\item{data}{Dataset with required quantities obtained from the
preprocessing function.}

\item{vessel_name}{Vessel name to plot trajectories as shown in the
vessel information database obtained from the get_vessel_info function.
'all' can be specified to plot trajectories of all fishing vessels.}
}
\value{
A data.frame with the global index of collocation (GIC) and
center of gravity.
}
\description{
Calculate global index of collocation between day i and day i+1
(if calculated for many vessels) or trip i and i+1 (if calculated
for a single vessel).
}
