% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Efficiency indicators}
\usage{
get_indicators(data)
}
\arguments{
\item{data}{Dataset with required quantities obtained from the
preprocessing function.}
}
\value{
A data.frame with the efficiency indicators per trip.
}
\description{
Calculate nine efficiency indicators.
}
