% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vessel_info.R
\name{get_vessel_info}
\alias{get_vessel_info}
\title{Vessel information}
\usage{
get_vessel_info(vmsdata)
}
\arguments{
\item{vmsdata}{VMS dataset.}
}
\value{
A data.frame with two columns: vessel name and code (matricula).
}
\description{
Get information for vessel name and code (matricula) in the VMS data.
}
