% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cgi.R
\name{plot_cgi}
\alias{plot_cgi}
\title{Plot inertia and center of gravity}
\usage{
plot_cgi(
  data,
  vessel_name,
  save_folder = "./",
  color_scale = c("blue", "white", "red"),
  alpha = 0.1,
  mainTitle = NULL,
  ...
)
}
\arguments{
\item{data}{Dataset with required quantities obtained from the
preprocessing function.}

\item{vessel_name}{Vessel name to plot trajectories as shown in the
vessel information database obtained from the get_vessel_info function.
'all' can be specified to plot trajectories of all fishing vessels.}

\item{save_folder}{Folder to save the plot.}

\item{color_scale}{Color scale to plot trip initial dates. Standard
scale specifies blue and red as the earliest and latest dates,
respectively.}

\item{alpha}{Alpha parameter for color scale.}

\item{mainTitle}{Main title in plot. Default value is the vessel name.}
}
\value{
A PNG where the center of gravity and intertia are shown.
Bubble size is proportinal to trip catch.
}
\description{
Make a plot with center of gravity per trip, trip catch, and trip
initial date.
}
