% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot trajectories}
\usage{
plot_trajectory(
  data,
  vessel_name = "all",
  save_folder = "./",
  save_plot = FALSE,
  ...
)
}
\arguments{
\item{data}{Dataset with required quantities obtained from the
preprocessing function.}

\item{vessel_name}{Vessel name to plot trajectories as shown in the
vessel information database obtained from the get_vessel_info function.
'all' can be specified to plot trajectories of all fishing vessels.}

\item{save_folder}{Folder to save the plots.}

\item{save_plot}{Logical value to indicate if plot is saved as PNG.}

\item{...}{Other artuments for the plot function.}

\item{vesseldata}{Vessel information.}
}
\value{
PNG files are created in the specified folder.
}
\description{
Make plots of trajectories per fishing trip.
}
