% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmsdata-data.R
\docType{data}
\name{vmsdata}
\alias{vmsdata}
\title{VMS data example}
\format{
An object of class \code{"data.frame"}
}
\source{
Private data
}
\usage{
data(vmsdata)
}
\description{
All datasets that contain VMS information should follow this format. Be
careful with column names, they should be exactly as shown in this example.
The column DC is optional and should be name DIST_COAST later on.
}
\examples{
data(vmsdata)
}
\references{
None
}
\keyword{datasets}
