% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicators.R
\name{plot_indicators}
\alias{plot_indicators}
\title{Plot efficiency indicators}
\usage{
plot_indicators(ind_data, save_folder = "./", ...)
}
\arguments{
\item{ind_data}{Dataset with efficiency indicators obtained from the
get_indicators function.}

\item{save_folder}{Folder to save the plot.}
}
\value{
A PNG with the efficiency indicators plot.
}
\description{
Make a plot of the nine efficiency indicators per period (YYYYMM).
}
